+++ db/migrate/20250911125234_add_source_to_raif_tasks.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddSourceToRaifTasks < ActiveRecord::Migration[7.1]
+  def change
+    add_reference :raif_tasks, :source, polymorphic: true, index: true
+  end
+end