+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added `before_prompt_model_for_entry_response` callback to `Raif::Conversation` subclasses. [#233](https://github.com/CultivateLabs/raif/pull/233)
 - Added `generating_entry_response` flag to `Raif::Conversation` to track when a conversation is generating an entry response. [#241](https://github.com/CultivateLabs/raif/pull/241)
 - Added support for using other OpenAI API providers (e.g. Azure) via `Raif.config.open_ai_base_url` and `Raif.config.open_ai_api_version`. [#245](https://github.com/CultivateLabs/raif/pull/245)
+- Added an optional `source` association to `Raif::Task` for tracking the source of a task. [#246](https://github.com/CultivateLabs/raif/pull/246)
 
 ## v1.3